+++ spec/support/admin.rb
@@ -1,0 +1,7 @@
+# -*- encoding : utf-8 -*-
+class Admin
+  include Mongoid::Document
+  include Mongoid::Userstamp::User
+
+  field :name
+end