+++ spec/support/book.rb
@@ -3,5 +3,7 @@ class Book
   include Mongoid::Document
   include Mongoid::Userstamp
 
+  mongoid_userstamp user_model: 'User'
+
   field :name
 end